Isolate legacy codec pipeline
